third_party/blink/web_tests/fast/css/content-language-dynamically-added-expected.txt

Test for bug 76701: map http-equiv content-language to -webkit-locale. This particular test tests that a dynamically added content-language is processed. This expectation may change, see bug. The HTML 5 spec decrees that http-equiv content-language be processed when it is inserted into the document. In Firefox and IE adding the meta element dynamically doesn't seem to affect the page, but in Firefox the meta content-language displayed in "Page Information" is affected.

PASS languageOfNode('x') is "auto"
PASS languageOfNode('y') is "\"ar\""
PASS languageOfNode('x') is "\"ja\""
PASS languageOfNode('y') is "\"ar\""
PASS successfullyParsed is true

TEST COMPLETE
